refactor(tanstack-start): address circular dep and minor code issues
- Extract TravelParty/partyLabel into lib/travel-party.ts to break circular dep between BundleCard and offers route
- Fix ageGroup label fallback (map lookup instead of hardcoded "anyone" key)
- Add entitlement fallback for user_profile labels without a known ageGroup
- Replace Array.fill({}) with Array.from to avoid shared-reference bugs
- Remove no-op spread in TravelerPicker setCount
- Add `as const` on entitlement type discriminator to satisfy strict TS
